+++ example/example.dart
@@ -1,27 +1,26 @@
-# Lite Switch
-Lite Switch is a customized switch which you can add text and icon for "On" and "Off".
-You can also set the width and height for the switch in your app.
-
-## Screenshot
- <img height="414" src="https://github.com/galacticpuffin/lite_switch/raw/master/assets/images/on.png">
- <img height="414" src="https://github.com/galacticpuffin/lite_switch/raw/master/assets/images/off.png">
-
-
-
-## Usage
-To use this plugin :
-```dart
-Add the dependency to your pubspec.yaml file.
-  dependencies:
-    flutter:
-      sdk: flutter
-      lite_switch:
-```
-
-## Example
-```dart
+import 'package:flutter/material.dart';
 import 'package:lite_switch/lite_switch.dart';
-Widget build(BuildContext context) {
+
+void main() => runApp(MyApp());
+
+class MyApp extends StatefulWidget {
+  @override
+  _LiteSwitchState createState() => new _LiteSwitchState();
+}
+
+class _LiteSwitchState extends State<MyApp> {
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
     return new Scaffold(
       body: new Center(
         child: new Column(
@@ -49,5 +48,5 @@ Widget build(BuildContext context) {
         ),
       ),
     );
- }
-```
+  }
+}