+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+## [1.0.4] Apri 14, 2020
+* Change class name using UpperCamelName
+* Update Package description
+* fix screenshot link bug
+* Update License
+* Add Example
+
+## [1.0.3] Apri 14, 2020
+
+* Fix screenshot link bug
+
 ## [1.0.2] Apri 14, 2020
 
-* fix screenshot link
+* Fix screenshot link
 
 ## [1.0.1] Apri 14, 2020